flatten algorithm structure
