--- plugin/algorithms/__init__.py
@@ -1,2 +1,0 @@
-from .superimpose_by_entry import *
-from .superimpose_by_chain import *
