+++ plugin/superimpose_by_entry.py
@@ -5,6 +5,7 @@ from nanome.api.structure import Complex
 
 from . import utils
 
+__all__ = ["superimpose_by_entry"]
 
 def superimpose_by_entry(fixed_comp: Complex, moving_comp: Complex, overlay_method):
     parser = PDBParser(QUIET=True)
